Add files via upload
In this project I had to make a clock program that had a 12 hour and 24 hour clock outputted on the screen with an options menu to add an hour, min, or sec. I feel like I did well at creating a clock program that output the necessary elements. I feel like I could have improved the code by adding a loop for the times and figuring out how to format the am and pm correctly. This would have helped with the length of the code and readability.